requirements.txt
seleniumbeautifulsoup4